+++ types/index.d.ts
@@ -16,6 +16,9 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
+import { BSONRegExp } from "bson";
+import { ObjectID } from "bson";
+
 // TypeScript Version: 2.3.2
 // With great contributions to @akim95 on github
 
@@ -435,6 +438,8 @@ declare namespace Realm {
         function initiateClientReset(path: string): void;
         function _hasExistingSessions(): boolean;
         function reconnect(): void;
+        function allSyncSessions(user: Realm.User): [Realm.App.Sync.Session];
+        function syncSession(user: Realm.User, partitionValue: string|number|ObjectID|null) : Realm.App.Sync.Session;
     }
 }